Various changes for "level of detail" rendering and modal toggling.
Modal toggling logic and event listeners moved to Modal itself, which
listens to Demo.state.levelOfDetail.

Demo contains a callback which adds/removes classes to HTML doc body
to represent the "level of detail" selected by the slider in the nav
menu.

Desk adds classes (.lod and .lod-*) to each sentence in each
paragraph. This is very inefficient but works for now!