Good lord, this was horrible.
Safari on iOS was showing a blank screen. The *only* error in the console was about an uncaught promise raised by Preact.

After a lot of digging, this was being caused by Safari not understanding some CSS for styling a range slider thumb.

So these styles have been removed, and the slider thumb is unstyled.

Also upgraded to the latest version of Preact as part of the debugging, and decided to keep it.

Safari on iOS is **EXTREMELY** unreliable when it comes to refreshing a page and discarding the cached version. This makes debugging a TRULY AWFUL experience.